Copied the files from the antalexco.ro site built in November 2009 as they contain the seed for a very lightweight cms (called at the time bitSimpleCMS).
Template background, logo, content text and other images have been removed and replaced with placeholder images and css styled divs in order to anonymize content and reduce size.

The CMS basically resides in two files, content_vars.php and content_funcs.php and they need to be included at the beginning of index.php before including template.html which in turn calls for various functions to generate menu and content (the functions are named simply GenerateMenu(), GenerateContent(), ...LangIcons(), ...Title() etc). The reason to include them is that it is desirable to think the CMS from a lightweight and independent perspective before integrating it with quarkOS as we want it to be able to function both by itself (in very light setups) and as a service on top of quarkOS.